+++ src/frontend/src/views/HomeView.vue
@@ -12,6 +12,7 @@ import ChatsSidebar from "@/components/ChatsSidebar";
 import getUserData from "@/hooks/getUserData";
 import getRelatedUsers from "@/hooks/getRelatedUsers"
 import { useStore } from "vuex";
+import refresh from "@/hooks/refresh"
 import {computed, onMounted, ref} from "vue";
 export default {
   name: 'HomeView',
@@ -58,14 +59,16 @@ export default {
                 store.dispatch('addUser', {user: user, chatId: chatId.value, chat: data.chat})
             }
         } else if (data.deleted_user) {
-            console.log(data.deleted_user)
-            store.dispatch('deleteUser', {user: data.deleted_user, chatId: chatId.value});
+            if (store.getters.getUserName !== data.deleted_user.username) {
+              store.dispatch('deleteUser', {user: data.deleted_user, chatId: chatId.value});
+            }
             if (store.getters.getUserName === data.deleted_user.username) {
                 store.commit('deleteChat', {chatId: chatId.value})
             }
         } else if (data.leaved_user) {
-            console.log(data.leaved_user)
-            store.dispatch('deleteUser', {user: data.leaved_user, chatId: chatId.value});
+            if (store.getters.getUserName !== data.leaved_user.username) {
+              store.dispatch('leaveUser', {user: data.leaved_user, chatId: chatId.value});
+            }
             if (store.getters.getUserName === data.leaved_user.username) {
                 store.commit('deleteChat', {chatId: chatId.value})
             }
@@ -75,6 +78,9 @@ export default {
 
     onMounted(async () => {
       fetchUserData();
+      setInterval(() => {
+        refresh();
+      }, 14 * 60 * 1000); 
     });
 
     return {userData}
